requirements.txt
flask==3.1.3
pandas==2.1.4
networkx==3.2.1
spacy==3.7.2
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
# مدل‌های جدید
transformers==4.36.2
torch==2.1.2
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1
# کتابخانه‌های مورد نیاز برای ارزیابی
nltk==3.8.1
sentence-transformers==2.2.2
textblob==0.17.1
jieba==0.42.1
json-repair==0.19.0 